Make participation lists actually work
